+++ Libraries/AQ_OSD/MAX7456_Config.h
@@ -39,7 +39,7 @@
 
 //Battery info - 5-16 characters long
 #define VOLTAGE_ROW 2
-#define VOLTAGE_COL 1
+#define VOLTAGE_COL 0
 
 //Compass reading - 5 characters long
 #define COMPASS_ROW 1
@@ -47,7 +47,7 @@
 
 //Altitude reading - up to 8 characters long (32768 max)
 #define ALTITUDE_ROW 1
-#define ALTITUDE_COL 1
+#define ALTITUDE_COL 0
 
 //Flight timer - 6 characters long
 #define TIMER_ROW 1
@@ -62,7 +62,9 @@
 
 // RSSI monitor
 #define RSSI_ROW     3
-#define RSSI_COL     23
+#define RSSI_COL     24
+#define SIGNAL_QUALITY_ROW 3
+#define SIGNAL_QUALITY_COL 0
 
 // GPS info
 #define GPS_ROW     MAX_screen_rows-2
